Split SwarmRunner
