+++ planet-clicker/Assets/LibplanetUnity/SwarmRunner.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0ec4f0cb45505a0489b3c113821a4539
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
